[MOB-15749] -- Move ExecuteGeminioTemplateAction into new package
